+++ ProyectoDWS/src/main/java/com/example/proyectodws/rest/CommentRestController.java
@@ -19,7 +19,6 @@ import java.sql.Date;
 import java.time.LocalDate;
 import java.util.List;
 
-// Rest controller for comments.
 @RestController
 @RequestMapping("/api/comments")
 public class CommentRestController {
@@ -33,7 +32,6 @@ public class CommentRestController {
     @Autowired
     private CourseService courseService;
 
-    // Get comments by course
     @GetMapping("/course/{courseId}")
     public ResponseEntity<CommentsResponse> getCommentsByCourse(@PathVariable Long courseId) {
         List<CommentWithIdsDTO> comments = commentService.getCommentsForCourseCompact(courseId);
@@ -49,24 +47,17 @@ public class CommentRestController {
         return ResponseEntity.ok(new CommentsResponse(new GenericResponse("Comentarios obtenidos correctamente", 200), comments));
     }
 
-    // Add comment
     @PostMapping
     public ResponseEntity<CommentResponse> createComment(
             @RequestParam Long courseId,
             @RequestParam String text
     ) {
-
-        // Validate text
         if (text == null || text.trim().isEmpty()) {
             return ResponseEntity.status(400).body(new CommentResponse(new GenericResponse("El texto es obligatorio", 400), null));
         }
-
-        // Validate courseId
         if (courseId == null || courseId < 0) {
             return ResponseEntity.status(400).body(new CommentResponse(new GenericResponse("El curso es obligatorio", 400), null));
         }
-
-        // Check if course exists
         CourseDTO course = courseService.getCourseById(courseId);
         if (course == null) {
             return ResponseEntity.status(404).body(new CommentResponse(new GenericResponse("Curso no encontrado", 404), null));
@@ -74,7 +65,6 @@ public class CommentRestController {
 
         UserDTO user = userService.getLoggedUserDTO();
 
-        // Create comment
         CommentDTO comment = new CommentDTO(
                 null,
                 text,
@@ -84,17 +74,25 @@ public class CommentRestController {
         return ResponseEntity.ok(new CommentResponse(new GenericResponse("Comentario creado correctamente", 200), commentService.addComment(user.id(), courseId, comment)));
     }
 
-    // Delete comment
     @DeleteMapping("/{commentId}")
     public ResponseEntity<GenericResponse> deleteComment(@PathVariable Long commentId) {
 
-        // Check if comment exists
         CommentDTO comment = commentService.getCommentById(commentId);
         if (comment == null) {
             return ResponseEntity.status(404).body(new GenericResponse("Comentario no encontrado", 404));
         }
 
-        // Delete comment
+        UserDTO commentUser = commentService.getUserFromCommentId(commentId);
+        if (commentUser == null) {
+            return ResponseEntity.status(404).body(new GenericResponse("Comentario no encontrado", 404));
+        }
+
+        UserDTO currentUser = userService.getLoggedUserDTO();
+        if (!currentUser.roles().contains("ADMIN") && !commentUser.id().equals(currentUser.id())) {
+            return ResponseEntity.status(403).body(new GenericResponse("No tienes permisos para eliminar este comentario", 403));
+        }
+
+
         commentService.deleteComment(commentId);
         return ResponseEntity.ok(new GenericResponse("Comentario eliminado correctamente", 200));
     }
